Fix for problem where if you kill Toszylae and walk away without the helmet, you are doomed. There is nothing in any subsequent dialogue to give you a hint to go and look for something where you killed the lich, and in fact Talion's dialogue tree is simply wrong. The only way to get past talion_demon_1 is to have darkprotector:10. If you don't have that, the only available reply is "No, I have not killed it yet", even though you have. The only thing you can do with Urlifendor is go through a long dialogue sequence that arrives at the selector ulifendor_infected_18_s. That has two replies, but one requires toszylae:70, which you can't get until you have darkprotector:10, so of course you don't have it. So the only available reply is "OK. I will defeat the lich first. Goobye", even though you have already defeated it.
The helmet is now on the shrine behind the lich. When you pick it up, it disappears (mapchange). If you try to leave without it, you get a very big hint.

This is still not perfect. I could not figure out any way to change Talion's or Ulifendor's dialogue without a huge amount of work (the logic in the dialogues and the three interleaved quests is very bad, and in some cases only works because one mistake cancels another. I do not want to mess with it). This means that if the player kills the lich and still leaves without the helmet, despite the hint, the dialogues still make no sense. That is much less likely to happen now though.

I have tested this, and everything seems to work as desired.

There are also a number of other typo fixes, as reported on the forums, and I made the GoW extraordinary (as also reported on the forums).